src/templates/template-vision-segmentation/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pytorch-ignite>=0.4.4
pyyaml
albumentations
image_dataset_viz

#::: if (['neptune', 'polyaxon'].includes(it.logger)) { :::#

#:::= it.logger + '-client' :::#

#::: } else { :::#

#:::= it.logger :::#

#::: } :::#
